+++ Core/IconSettingsRenderer.lua
@@ -456,13 +456,6 @@ function IconSettingsRenderer:GetIconConfigInputs(config)
                 },
                 {
                     type = "checkbox",
-                    label = "Enable spell charges cooldown tracking",
-                    tooltip = "This feature is experimental. Enabling will track the cooldown for each charge spent. This will cause the global cooldown to also apply on the Icon. The status bar will still be able to ignore the GCD. If you leave this disabled, the spell cooldown will only fire when the last charge of the spell is used. If you have 1 or more charges, it will behave like other spells that are 'available' to cast. Text for charges is unaffected.",
-                    getValue = function(self) return config.getValue(self.uniqueID, "iconSettings.trackIndividualChargeCooldown") or false end,
-                    setValue = function(self, value) config.setValue(self.uniqueID, "iconSettings.trackIndividualChargeCooldown", value) end,
-                },
-                {
-                    type = "checkbox",
                     label = "Desaturate when on cooldown",
                     getValue = function(self) return config.getValue(self.uniqueID, "iconSettings.desaturated") or false end,
                     setValue = function(self, value) config.setValue(self.uniqueID, "iconSettings.desaturated", value) end,
@@ -472,13 +465,6 @@ function IconSettingsRenderer:GetIconConfigInputs(config)
                     label = "Hide default swipe animation",
                     getValue = function(self) return config.getValue(self.uniqueID, "iconSettings.hideDefaultSweep") == true end,
                     setValue = function(self, value) config.setValue(self.uniqueID, "iconSettings.hideDefaultSweep", value) end,
-                },
-                {
-                    type = "checkbox",
-                    label = "Is Channeled Spell",
-                    tooltip = "Channeled spells must have an indication that they are channeled in order to properly respond to the cooldown events. Unchecked can cause them to only respond to the GCD.",
-                    getValue = function(self) return config.getValue(self.uniqueID, "iconSettings.isChanneledSpell") or false end,
-                    setValue = function(self, value) config.setValue(self.uniqueID, "iconSettings.isChanneledSpell", value) end,
                 },
                 {
                     type = "checkbox",